PdfGen/requirements.txt
docx2pdf
fpdf2
PyPDF2
arabic-reshaper
reportlab
